chore: forge fmt

+++ src/borrower/deployer.sol
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.7.6;
 
-import { ShelfFabLike, PileFabLike, TitleFabLike } from "./fabs/interfaces.sol";
-import { FixedPoint } from "./../fixed_point.sol";
+import {ShelfFabLike, PileFabLike, TitleFabLike} from "./fabs/interfaces.sol";
+import {FixedPoint} from "./../fixed_point.sol";
 
 interface DependLike {
     function depend(bytes32, address) external;
@@ -18,20 +18,20 @@ interface NAVFeedLike {
 }
 
 interface FeedFabLike {
-    function newFeed() external returns(address);
+    function newFeed() external returns (address);
 }
 
 interface FileLike {
-    function file(bytes32 name, uint value) external;
+    function file(bytes32 name, uint256 value) external;
 }
 
 contract BorrowerDeployer is FixedPoint {
-    address      public immutable root;
+    address public immutable root;
 
-    TitleFabLike     public immutable titlefab;
-    ShelfFabLike     public immutable shelffab;
-    PileFabLike      public immutable pilefab;
-    FeedFabLike      public immutable feedFab;
+    TitleFabLike public immutable titlefab;
+    ShelfFabLike public immutable shelffab;
+    PileFabLike public immutable pilefab;
+    FeedFabLike public immutable feedFab;
 
     address public title;
     address public shelf;
@@ -39,23 +39,23 @@ contract BorrowerDeployer is FixedPoint {
     address public immutable currency;
     address public feed;
 
-    string  public titleName;
-    string  public titleSymbol;
+    string public titleName;
+    string public titleSymbol;
     Fixed27 public discountRate;
 
     address constant ZERO = address(0);
     bool public wired;
 
-    constructor (
-      address root_,
-      address titlefab_,
-      address shelffab_,
-      address pilefab_,
-      address feedFab_,
-      address currency_,
-      string memory titleName_,
-      string memory titleSymbol_,
-      uint discountRate_
+    constructor(
+        address root_,
+        address titlefab_,
+        address shelffab_,
+        address pilefab_,
+        address feedFab_,
+        address currency_,
+        string memory titleName_,
+        string memory titleSymbol_,
+        uint256 discountRate_
     ) {
         root = root_;
 
@@ -115,7 +115,7 @@ contract BorrowerDeployer is FixedPoint {
 
         AuthLike(feed).rely(shelf);
         AuthLike(title).rely(shelf);
-        
+
         FileLike(feed).file("discountRate", discountRate.value);
 
         if (initNAVFeed) {
@@ -123,9 +123,7 @@ contract BorrowerDeployer is FixedPoint {
         }
     }
 
-
     function deploy() public {
         deploy(false);
     }
 }
-
