+++ test/borrower/mock/feed.sol
@@ -1,5 +1,6 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.7.6;
+
 import "ds-test/test.sol";
 
 import "../../../test/mock/mock.sol";
@@ -10,19 +11,19 @@ contract NAVFeedMock is Mock, Auth {
         wards[msg.sender] = 1;
     }
 
-    function pile() public view returns(address){ 
-            return values_address_return["pile"];
+    function pile() public view returns (address) {
+        return values_address_return["pile"];
     }
 
-    function shelf() public view returns(address){ 
-            return values_address_return["shelf"];
+    function shelf() public view returns (address) {
+        return values_address_return["shelf"];
     }
 
-    function maturityDate(bytes32 nft_) public view returns(uint){ 
-            return values_return["maturityDate"];
+    function maturityDate(bytes32 nft_) public view returns (uint256) {
+        return values_return["maturityDate"];
     }
 
-    function nftID(uint loan) public returns (bytes32) {
+    function nftID(uint256 loan) public returns (bytes32) {
         return values_bytes32_return["nftID"];
     }
 }
