+++ test/lender/mock/memberlist.sol
@@ -6,23 +6,22 @@ import "tinlake-auth/auth.sol";
 import "../../../test/mock/mock.sol";
 
 contract MemberlistMock is Mock, Auth {
-
     constructor() {
         wards[msg.sender] = 1;
     }
 
     function updateMember(address usr, uint256 validUntil) public auth {
-      calls["updateMember"]++;
-      values_address["updateMember_usr"] = usr;
-      values_uint["updateMember_validUntil"] = validUntil;
+        calls["updateMember"]++;
+        values_address["updateMember_usr"] = usr;
+        values_uint["updateMember_validUntil"] = validUntil;
     }
 
     function updateMembers(address[] memory users, uint256 validUntil) public auth {
-      calls["updateMembers"]++;
+        calls["updateMembers"]++;
 
-      for (uint i=0; i<users.length; i++) {
-        values_address["updateMembers_usr"] = users[i];
-        values_uint["updateMembers_validUntil"] = validUntil;
-      }
+        for (uint256 i = 0; i < users.length; i++) {
+            values_address["updateMembers_usr"] = users[i];
+            values_uint["updateMembers_validUntil"] = validUntil;
+        }
     }
 }
