+++ test/lender/mock/pool.sol
@@ -1,11 +1,12 @@
 // SPDX-License-Identifier: AGPL-3.0-only
 pragma solidity >=0.6.12;
+
 import "forge-std/Test.sol";
 
 import "../../../test/mock/mock.sol";
 
 contract PoolMock is Mock {
-    function totalValue() public view returns (uint) {
+    function totalValue() public view returns (uint256) {
         return values_return["totalValue"];
     }
 }
